+++ src/ze-grey-client-test.c
@@ -177,7 +177,7 @@ do_test(conf_T * cargs)
     if (cargs->random) {
       rind = random() % range;
 
-      snprintf(ipbuf, sizeof (ipbuf), "1.1.%d.%d", rind % 255, random() % 255);
+      snprintf(ipbuf, sizeof (ipbuf), "1.1.%d.%d", (int ) rind % 255, (int ) (random() % 255));
       ip = ipbuf;
     } else
       rind = n;
